requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.340
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
openai==0.28.0
pymilvus==2.3.3
transformers==4.35.2
torch==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
tiktoken==0.5.1
rank-bm25==0.2.2
streamlit-chat==0.1.0
plotly==5.18.0
faiss-cpu==1.7.4
pypdf==3.17.0
python-dotenv==1.0.0
onnxruntime==1.16.0